pyproject.toml
[project]
name = "streamlit-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-core>=0.14.4",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-groq>=0.4.1",
    "llama-index-readers-file>=0.5.4",
    "streamlit>=1.50.0",
]
